FIX: user layout
