+++ src/components/layout/ContainerWrapper.vue
@@ -2,7 +2,7 @@
 
 <template>
   <div class="container-fluid m-0 p-0 min-vh-100 text-bg-gray-100">
-    <div class="container-xxl min-vh-100 m-auto d-flex justify-content-center align-items-center">
+    <div class="container-xxl m-auto d-flex justify-content-center align-items-center">
       <slot name="content" />
     </div>
   </div>
